requirements.txt
blinker==1.4
dominate==2.1.17
Flask==0.10.1
Flask-Bootstrap==3.3.5.7
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pyserial==3.0.1
python-dateutil==2.5.0
six==1.10.0
SQLAlchemy==1.0.12
visitor==0.1.2
Werkzeug==0.11.4
